+++ artemis/experiments/experiment_record_view.py
@@ -5,6 +5,7 @@ from artemis.experiments.experiment_management import load_lastest_experiment_re
 from artemis.experiments.experiment_record import NoSavedResultError, ExpInfoFields, ExperimentRecord
 from artemis.experiments.experiments import is_experiment_loadable, GLOBAL_EXPERIMENT_LIBRARY
 from artemis.general.display import deepstr, truncate_string, hold_numpy_printoptions, side_by_side
+from artemis.general.nested_structures import flatten_nested_object
 from artemis.general.should_be_builtins import separate_common_items, all_equal, bad_value, izip_equal
 from artemis.general.tables import build_table
 from tabulate import tabulate
@@ -78,10 +79,8 @@ def get_record_invalid_arg_string(record):
         current_args = dict(record.get_experiment().get_args())
         validity = record.is_valid(last_run_args=last_run_args, current_args=current_args)
         if validity is False:
-            last_arg_str, this_arg_str = [
-                ['{}:{}'.format(k, v) for k, v in argdict.iteritems()] if isinstance(argdict, dict) else
-                ['{}:{}'.format(k, v) for k, v in argdict]
-                for argdict in (last_run_args, current_args)]
+            last_arg_str = stringify_arguments(last_run_args)
+            this_arg_str = stringify_arguments(current_args)
             common, (old_args, new_args) = separate_common_items([last_arg_str, this_arg_str])
             notes = "No: Args changed!: {{{}}}->{{{}}}".format(','.join(old_args), ','.join(new_args))
         elif validity is None:
@@ -92,6 +91,25 @@ def get_record_invalid_arg_string(record):
         notes = "<Experiment Not Currently Imported>"
     return notes
 
+
+def stringify_arguments(args):
+    '''
+    Returns a list of strings of a dict, such that each list entry has the form 'k:v'. If a value of the dict is itself a dict, the nested structure is
+    unpacked such that the returned list has elements [k1:k11:a, k1:k12:b] etc.
+    :param args:
+    :return:
+    '''
+    def recursively_parse_dict(d):
+        res = []
+        for k,v in d.iteritems():
+            if isinstance(v,dict):
+                sub_list = recursively_parse_dict(v)
+            else:
+                sub_list = [v]
+            res.extend(['{}:{}'.format(k,s) for s in sub_list])
+        return res
+    res = recursively_parse_dict(args)
+    return res
 
 def get_oneline_result_string(record, truncate_to=None, array_float_format='.3g', array_print_threshold=8):
     """
@@ -283,3 +301,4 @@ def make_record_comparison_table(records, args_to_show=None, results_extractor =
         import tabulate
         print tabulate.tabulate(rows, headers=headers, tablefmt='simple')
     return headers, rows
+
